refactor: handle conflicts
