+++ gatsby-config.js
@@ -10,5 +10,12 @@ module.exports = {
         redirect: true,
       },
     },
+    {
+      resolve: `gatsby-source-contentful`,
+      options: {
+        spaceId: 'nnjxq94o0k8q',
+        accessToken: 'arJoI5VPa5lumtee5XAHhnZSnRK9dsuM_dDHoi8IFks',
+      },
+    },
   ]
 }
