+++ package.json
@@ -17,6 +17,7 @@
     "gatsby": "^2.13.42",
     "gatsby-plugin-intl": "^0.2.7",
     "gatsby-plugin-sass": "^2.1.3",
+    "gatsby-source-contentful": "^2.1.16",
     "jquery": "^3.4.1",
     "node-sass": "^4.12.0",
     "popper.js": "^1.15.0",
