+++ src/components/layout.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './header/header';
+import Header from './header';
 import Footer from './footer';
 
 const Layout = (props) => {
