+++ src/styles/index.scss
@@ -1,4 +1,4 @@
-@import '../components/header/header.scss';
+@import './header.scss';
 @import './director-of-day.scss';
 @import './description-home-page.scss';
